Create README.md
Q. What is java?
Answer - Java was developed by James Gosling at Sun Microsystems in 1991. It was first named Oak afterward they changed to java because they wanted a unique & name that reflects the essence of technology. Java is an island where the first coffee was produced.
It is an object-oriented programming language.